adjust props
